refactor(seeder): address review feedback on provider scenes
- Collapse the inline ProviderUser block in SingleProviderScene to
  ProviderUserSeeder.CreateConfirmedAdmin so the recipe lives in one place.
- Inline the single-use ProviderSceneHelpers.ResolveDomain into the scene and
  remove the helper class.
- Rename SeederBilling to BillingEmailSeeder for consistency with sibling
  factories (and the matching tests).
- Move wrap_symmetric_key out of cipher.rs into a dedicated provider.rs module
  (key wrapping is not cipher encryption); register it in build.rs and lib.rs.
- Move the provider scene integration test to
  Scenes/SingleProviderSceneTests.cs.
